requirements.txt
ffmpeg-python==0.2.0
openai==1.2.0